+++ README.md
@@ -10,7 +10,7 @@ Implement the `App` with 2 pages and ability to select a person in the table.
     - add `replace` attribute not to save `/home` URL in the [browser history](https://reactrouter.com/en/main/start/tutorial#managing-the-history-stack) and avoid navigation loop when you press browser `Go back` button.
 1. Implement `NotFoundPage` with a title `Page not found` that is shown for all the other URLs;
 1. Add the `Navbar` with 2 links `Home` and `People`. Active one should have the `has-background-grey-lighter` class;
-1. Use `HashRouter` to be able to share a link to any page on Github pages; 
+1. Use `HashRouter` to be able to share a link to any page on Github pages;
 1. Fetch `people` from [the API](https://mate-academy.github.io/react_people-table/api/people.json) when `PeoplePage` is opened;
 1. Pass `people` to `PeopleTable` component and render as a table (see the given markup);
 1. Make each name in the table a link a to `/people/:slug` (including mother and father)
@@ -25,4 +25,4 @@ Implement the `App` with 2 pages and ability to select a person in the table.
 - Implement a solution following the [React task guideline](https://github.com/mate-academy/react_task-guideline#react-tasks-guideline).
 - Use the [React TypeScript cheat sheet](https://mate-academy.github.io/fe-program/js/extra/react-typescript).
 - Open one more terminal and run tests with `npm test` to ensure your solution is correct.
-- Replace `<your_account>` with your Github username in the [DEMO LINK](https://<your_account>.github.io/react_people-table-basics/) and add it to the PR description.
+- Replace `<your_account>` with your Github username in the [DEMO LINK](https://Maks-Tsarenko.github.io/react_people-table-basics/) and add it to the PR description.
