+++ src/components/PeopleList.tsx
@@ -1,25 +1,7 @@
 import classNames from 'classnames';
 import { Link } from 'react-router-dom';
 import { Person } from '../types';
-
-const findParent = (people: Person[], parentName: string | null) => {
-  const parent = people.find(person => person.name === parentName);
-
-  if (parent) {
-    return (
-      <Link
-        className={classNames({
-          'has-text-danger': parent.sex === 'f',
-        })}
-        to={`/people/${parent.slug}`}
-      >
-        {parent.name}
-      </Link>
-    );
-  }
-
-  return parentName || '-';
-};
+import { findParent } from '../utils/findParents';
 
 export type Props = {
   people: Person[],
