+++ src/pages/PageNotFound.tsx
@@ -1,0 +1,3 @@
+export const PageNotFound = () => (
+  <h1 className="title">Page not found</h1>
+);
